Restructuring overlays a bit
Editor is now a svelte only component. Screens however will be agnostic.
