+++ package/lib/overlay/editor/Inspector.svelte
