+++ package/lib/overlay/index.ts
@@ -1,5 +1,3 @@
-import { DotsScreen } from "./components/screen/web";
-import { DotsEditor } from "./components/editor/web";
+import { DotsScreen } from "./screen/web";
 
 customElements.define("dots-screen", DotsScreen);
-customElements.define("dots-editor", DotsEditor);
