+++ package/lib/overlay/types.d.ts
